requirements.txt
Django==5.0.0
Pillow==10.1.0
django-cors-headers==4.3.1
python-decouple==3.8
numpy==1.26.2
opencv-python==4.8.1.78
google-generativeai==0.3.2
